restructure containers and components into directories
