+++ src/components/ChartIcon/index.tsx
@@ -1,0 +1,3 @@
+import ChartIcon from './ChartIcon';
+
+export default ChartIcon;
