+++ src/components/ChartList/ChartList.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import './ChartList.css';
-import { ChartState } from '../shared/types';
+import { ChartState } from '../../shared/types';
 import ChartListItem from './ChartListItem';
 
 interface Props {
