+++ src/components/ChartList/index.tsx
@@ -1,0 +1,3 @@
+import ChartList from './ChartList';
+
+export default ChartList;
