+++ src/components/ChartView/index.tsx
@@ -1,0 +1,3 @@
+import ChartView from './ChartView';
+
+export default ChartView;
