+++ src/components/Footer/index.tsx
@@ -1,0 +1,3 @@
+import Footer from './Footer';
+
+export default Footer;
