+++ src/components/Sidebar/index.tsx
@@ -1,0 +1,3 @@
+import Sidebar from './Sidebar';
+
+export default Sidebar;
