+++ src/containers/ChartListContainer/index.tsx
@@ -1,0 +1,3 @@
+import ChartList from './ChartListContainer';
+
+export default ChartList;
