+++ src/containers/ChartViewContainer/index.tsx
@@ -1,0 +1,3 @@
+import ChartView from './ChartViewContainer';
+
+export default ChartView;
